+++ .projectile
@@ -1,0 +1,2 @@
+-build
+-profiler.log
