+++ src/lib/board.lua
@@ -1,344 +1,5 @@
---- ユーザまたは QPU が操作するカーソルのクラス
-cursor_class = new_class()
-
-function cursor_class._init(_ENV)
-  init(_ENV)
-end
-
---- カーソルの初期化
-function cursor_class.init(_ENV)
-  x, y, _tick = 3, 6, 0
-end
-
---- カーソルを左に移動
-function cursor_class.move_left(_ENV)
-  if x > 1 then
-    x = x - 1
-  end
-end
-
---- カーソルを右に移動
-function cursor_class.move_right(_ENV, cols)
-  if x < cols - 1 then
-    x = x + 1
-  end
-end
-
---- カーソルを上に移動
-function cursor_class.move_up(_ENV, rows)
-  if y < rows then
-    y = y + 1
-  end
-end
-
---- カーソルを下に移動
-function cursor_class.move_down(_ENV)
-  if y > 1 then
-    y = y - 1
-  end
-end
-
---- カーソルの状態を更新
-function cursor_class.update(_ENV)
-  _tick = (_tick + 1) % 28
-end
-
---- カーソルを描画
-function cursor_class.render(_ENV, screen_x, screen_y)
-  if _tick < 14 then
-    sspr(32, 32, 19, 11, screen_x - 2, screen_y - 2)
-  else
-    sspr(56, 32, 21, 13, screen_x - 3, screen_y - 3)
-  end
-end
-
---- ブロック (量子ゲート) クラス
-block_class = new_class()
-block_class.block_match_animation_frame_count = 45
-block_class.block_match_delay_per_block = 8
-block_class.block_swap_animation_frame_count = 3
-block_class.sprites = transform({
-  -- default|landing|match
-  h = "0|1,1,1,2,2,2,3,3,1,1,1,1|24,24,24,25,25,25,24,24,24,26,26,26,0,0,0,27",
-  x = "16|17,17,17,18,18,18,19,19,17,17,17,17|40,40,40,41,41,41,40,40,40,42,42,42,16,16,16,43",
-  y = "32|33,33,33,34,34,34,35,35,33,33,33,33|56,56,56,57,57,57,56,56,56,58,58,58,32,32,32,59",
-  z = "48|49,49,49,50,50,50,51,51,49,49,49,49|12,12,12,13,13,13,12,12,12,14,14,14,48,48,48,15",
-  s = "4|5,5,5,6,6,6,7,7,5,5,5,5|28,28,28,29,29,29,28,28,28,30,30,30,4,4,4,31",
-  t = "20|21,21,21,22,22,22,23,23,21,21,21,21|44,44,44,45,45,45,44,44,44,46,46,46,20,20,20,47",
-  control = "36|37,37,37,38,38,38,39,39,37,37,37,37|60,60,60,61,61,61,60,60,60,62,62,62,36,36,36,63",
-  cnot_x = "52|53,53,53,54,54,54,55,55,53,53,53,53|64,64,64,65,65,65,64,64,64,66,66,66,52,52,52,67",
-  swap = "8|9,9,9,10,10,10,11,11,9,9,9,9|80,80,80,81,81,81,80,80,80,82,82,82,8,8,8,83",
-  ["?"] = "98|98,98,98,98,98,98,98,98,98,98,98,98|98,98,98,98,98,98,98,98,98,98,98,98,98,98,98,98",
-  ["#"] = "113|113,113,113,113,113,113,113,113,113,113,113,113|113,113,113,113,113,113,113,113,113,113,113,113,113,113,113,113"
-}, function(each)
-  local default, landing, match = unpack(split(each, "|"))
-  return {
-    default = default,
-    landing = split(landing),
-    match = split(match)
-  }
-end)
-
-function block_class._init(_ENV, _type, _span, _height)
-  type, sprite_set, span, height, _state, _timer_landing =
-      _type, sprites[_type], _span or 1, _height or 1, "idle", 0
-end
-
---- 状態が idle かどうかを返す
-function block_class:is_idle()
-  return self._state == "idle"
-end
-
-function block_class:is_hover()
-  return self._state == "hover"
-end
-
-function block_class.is_fallable(_ENV)
-  return not (type == "i" or type == "?" or is_swapping(_ENV) or is_freeze(_ENV) or is_match(_ENV))
-end
-
-function block_class:is_falling()
-  return self._state == "falling"
-end
-
-function block_class.is_reducible(_ENV)
-  return type ~= "i" and type ~= "?" and is_idle(_ENV)
-end
-
-function block_class:is_match()
-  return self._state == "match"
-end
-
--- おじゃまブロックが小さいブロックに分解した後の硬直中かどうか
-function block_class:is_freeze()
-  return self._state == "freeze"
-end
-
-function block_class:is_swapping()
-  return self:_is_swapping_with_right() or self:_is_swapping_with_left()
-end
-
-function block_class:_is_swapping_with_left()
-  return self._state == "swapping_with_left"
-end
-
-function block_class:_is_swapping_with_right()
-  return self._state == "swapping_with_right"
-end
-
-function block_class:is_swappable_state()
-  return self:is_idle() or self:is_falling()
-end
-
-function block_class:is_empty()
-  return self.type == "i" and not self:is_swapping()
-end
-
-function block_class.is_single_block(_ENV)
-  return type == 'h' or type == 'x' or type == 'y' or type == 'z' or type == 's' or type == 't'
-end
-
-function block_class:swap_with(direction)
-  self.chain_id = nil
-  self:change_state("swapping_with_" .. direction)
-end
-
-function block_class:hover(timer)
-  self.timer = timer or 12
-  self:change_state("hover")
-end
-
-function block_class:fall()
-  --#if assert
-  assert(self:is_fallable(), "block " .. self.type .. "(" .. self.x .. ", " .. self.y .. ")")
-  --#endif
-
-  if self:is_falling() then
-    return
-  end
-
-  self:change_state("falling")
-end
-
-function block_class.replace_with(_ENV, other, match_index, _chain_id, garbage_span, garbage_height)
-  new_block, _match_index, _tick_match, chain_id, other.chain_id, _garbage_span, _garbage_height =
-      other, match_index or 0, 1, _chain_id, _chain_id, garbage_span, garbage_height
-
-  change_state(_ENV, "match")
-end
-
-function block_class.update(_ENV)
-  if is_idle(_ENV) then
-    if _timer_landing > 0 then
-      _timer_landing = _timer_landing - 1
-    end
-  elseif is_swapping(_ENV) then
-    if _tick_swap < block_swap_animation_frame_count then
-      _tick_swap = _tick_swap + 1
-    else
-      chain_id = nil
-      change_state(_ENV, "idle")
-    end
-  elseif is_hover(_ENV) then
-    if timer > 0 then
-      timer = timer - 1
-    else
-      change_state(_ENV, "idle")
-    end
-  elseif is_match(_ENV) then
-    if _tick_match <= block_match_animation_frame_count + _match_index * block_match_delay_per_block then
-      _tick_match = _tick_match + 1
-    else
-      change_state(_ENV, "idle")
-
-      if _garbage_span then
-        new_block._tick_freeze = 0
-        new_block._freeze_frame_count = (_garbage_span * _garbage_height - _match_index) * block_match_delay_per_block
-        new_block:change_state("freeze")
-      end
-    end
-  elseif is_freeze(_ENV) then
-    if _tick_freeze < _freeze_frame_count then
-      _tick_freeze = _tick_freeze + 1
-    else
-      change_state(_ENV, "idle")
-    end
-  end
-end
-
-function block_class:render(screen_x, screen_y, screen_other_x)
-  local shake_dx, shake_dy, swap_screen_dx, sprite = 0, 0
-
-  do
-    local _ENV = self
-
-    if type == "i" then
-      return
-    end
-
-    swap_screen_dx = (_tick_swap or 0) * (8 / block_swap_animation_frame_count)
-    if _is_swapping_with_left(_ENV) then
-      swap_screen_dx = -swap_screen_dx
-    end
-
-    if is_idle(_ENV) and _timer_landing > 0 then
-      sprite = sprite_set.landing[_timer_landing]
-    elseif is_match(_ENV) then
-      local sequence = sprite_set.match
-      sprite = _tick_match <= block_match_delay_per_block and sequence[_tick_match] or sequence[#sequence]
-    elseif _state == "over" then
-      sprite = sprite_set.match[#sprite_set.match]
-    else
-      sprite = sprite_set.default
-    end
-
-    -- CNOT または SWAP の接続を描画
-    if other_x and x < other_x then
-      line(
-        screen_x + 3,
-        screen_y + 3,
-        screen_other_x + 3,
-        screen_y + 3,
-        is_match(_ENV) and 13 or 10
-      )
-    end
-  end
-
-  if self.type == "?" then
-    palt(0, false)
-    pal(13, self.body_color)
-  end
-
-  if self._state == "over" then
-    shake_dx, shake_dy = rnd(2) - 1, rnd(2) - 1
-    pal(6, 9)
-    pal(7, 1)
-  end
-
-  spr(sprite, screen_x + swap_screen_dx + shake_dx, screen_y + shake_dy)
-
-  palt(0, true)
-  pal(13, 13)
-  pal(6, 6)
-  pal(7, 7)
-end
-
-function block_class:attach(observer)
-  self.observer = observer
-end
-
-function block_class.change_state(_ENV, new_state)
-  _timer_landing, _tick_swap =
-      is_falling(_ENV) and 12 or 0, 0
-
-  local old_state = _state
-  _state = new_state
-
-  observer:observable_update(_ENV, old_state)
-end
-
---#if debug
-local type_string = {
-  i = '_',
-  control = '●',
-  cnot_x = '+',
-  swap = 'X'
-}
-
-local state_string = {
-  idle = " ",
-  swapping_with_left = "<",
-  swapping_with_right = ">",
-  hover = "^",
-  falling = "|",
-  match = "*",
-  freeze = "f",
-}
-
-function block_class:_tostring()
-  return (type_string[self.type] or self.type:upper()) .. state_string[self._state]
-end
-
---#endif
-
---- おじゃまブロック
-
-local garbage_block_colors = { 2, 3, 4 }
-local inner_border_colors = { nil, 14, 11, 9 }
-
---- 新しいおじゃまブロックを作る
-function garbage_block(_span, _height, _color, _chain_id, _tick_fall)
-  local garbage = setmetatable({
-    body_color = _color or garbage_block_colors[ceil_rnd(#garbage_block_colors)],
-    chain_id = _chain_id,
-    tick_fall = _tick_fall,
-    dy = 0,
-    first_drop = true,
-    _render_box = draw_rounded_box,
-    render = function(_ENV, screen_x, screen_y)
-      local y0, x1, y1, _body_color =
-          screen_y + (1 - height) * 8,
-          screen_x + span * 8 - 2,
-          screen_y + 6,
-          _state ~= "over" and body_color or 9
-
-      _render_box(screen_x, y0 + 1, x1, y1 + 1, 5)                                                    -- 影
-      _render_box(screen_x, y0, x1, y1, _body_color, _body_color)                                     -- 本体
-      _render_box(screen_x + 1, y0 + 1, x1 - 1, y1 - 1, _state ~= "over" and inner_border_color or 1) -- 内側の線
-    end
-  }, { __index = block_class("g", _span or 6, _height) })
-
-  --#if assert
-  assert(garbage.body_color == 2 or garbage.body_color == 3 or garbage.body_color == 4,
-    "invalid color: " .. garbage.body_color)
-  assert(2 < garbage.span, "span must be greater than 2")
-  assert(garbage.span < 7, "span must be less than 7")
-  --#endif
-
-  garbage.inner_border_color = inner_border_colors[garbage.body_color]
-
-  return garbage
-end
+require("lib/cursor")
+require("lib/block")
 
 --- ブロックのマッチパターン
 reduction_rules = transform(
