+++ test/swap_block_utest.lua
@@ -1,5 +1,6 @@
 require("engine/test/bustedhelper")
 require("test/test_helper")
+require("lib/block")
 
 describe('swap_block', function()
   local swap
